+++ modules/azure-databricks-workspace/outputs.tf
@@ -1,0 +1,9 @@
+output "workspace_name" {
+  value       = azurerm_databricks_workspace.module-databricks.name
+  description = "Name of the Databricks workspace"
+}
+
+output "workspace_id" {
+  value       = azurerm_databricks_workspace.module-databricks.id
+  description = "ID of the Databricks workspace"
+}