+++ outputs.tf
@@ -1,12 +1,24 @@
 output "workspace_name" {
-  value       = azurerm_databricks_workspace.module-databricks.name
-  description = "Name of Databricks workspace"
+  value       = module.databricks-workspace.workspace_name
+  description = "Name of the Databricks workspace"
 }
-output "security_group_name" {
-  value       = azurerm_network_security_group.databricks-subnet-sg.name
-  description = "Name of security group assigned to subnets"
+output "workspace_id" {
+  value       = module.databricks-workspace.workspace_id
+  description = "ID of the Databricks workspace"
 }
-output "security_group_id" {
-  value       = azurerm_network_security_group.databricks-subnet-sg.id
-  description = "ID of security group assigned to subnets"
+output "security_group_private_name" {
+  value       = module.security-groups.security_group_private_name
+  description = "Name of security group assigned to the private subnet"
+}
+output "security_group_private_id" {
+  value       = module.security-groups.security_group_private_id
+  description = "ID of security group assigned to the private subnet"
+}
+output "security_group_public_name" {
+  value       = module.security-groups.security_group_public_name
+  description = "Name of security group assigned to the public subnet"
+}
+output "security_group_public_id" {
+  value       = module.security-groups.security_group_public_id
+  description = "ID of security group assigned to the public subnet"
 }